소스 정리

+++ Src/mpu6050_dmp.c
@@ -46,7 +46,7 @@ float Pitch, Roll, Roll_reverse, Yaw, Rangle=0.0f, Pangle=0.0f;
 
 float base_pitch=0.0f, base_roll=0.0f, base_yaw=0.0f, base_roll_reverse=0.0f;
 float dqw=1.0f, dqx=0.0f, dqy=0.0f, dqz=0.0f, sign=0.0f;
-float ledPos = 0.0f;
+uint8_t ledPos = 0;
 float targetLedPos = 0;
 float targetAnglel = 120.0f;
 uint8_t Cal_done = 0;
